+++ test/ts/default-import-as/src/index.ts
@@ -1,0 +1,4 @@
+import * as ansis from 'ansis';
+
+const message: string = `Hello, ${ansis.red`World`}!`;
+console.log(message);