+++ test/ts/named-import-color-func/src/index.ts
@@ -1,0 +1,4 @@
+import { red, ansi256 } from 'ansis';
+
+const message: string = `${ansi256(165)`Hello`}, ${red`World`}!`;
+console.log(message);