+++ test/ts/named-import-simple/src/index.ts
@@ -1,0 +1,4 @@
+import { red } from 'ansis';
+
+const message: string = `Hello, ${red`World`}!`;
+console.log(message);