+++ vitest.config.js
@@ -1,6 +1,4 @@
-import { defineConfig } from 'vitest/config';
-
-export default defineConfig({
+export default {
   test: {
     include: [
       'test/**/*.test.js',
@@ -10,5 +8,12 @@ export default defineConfig({
         'src/**/*',
       ],
     },
+    server: {
+      deps: {
+        inline: [
+          'AssertionError: ts',
+        ]
+      }
+    },
   },
-});
+}